+++ packages/display/src/components/Display.vue
@@ -1,7 +1,19 @@
 <template>
-  <VForm ref="form" class="tce-root" @submit.prevent="submit">
-    <div class="px-2 my-4">{{ data.question }}</div>
-    <div v-for="premise in data.premises" :key="premise.key" class="px-2 my-2">
+  <QuestionContainer
+    :data="data"
+    :is-correct="userState.isCorrect"
+    :is-submitted="isSubmitted"
+    allowed-retake
+    is-graded
+    @retry="isSubmitted = false"
+    @submit="submit"
+  >
+    <div class="text-subtitle-2 mb-2">Select correct answer for each:</div>
+    <div
+      v-for="premise in data.premises"
+      :key="premise.key"
+      class="text-subtitle-2 mt-2"
+    >
       <div class="mb-2">
         <span class="font-weight-bold">{{ data.headings.premise }}:</span>
         {{ premise.value }}
@@ -9,54 +21,40 @@
       <VSelect
         :items="data.responses"
         :label="data.headings.response"
-        :readonly="submitted"
+        :readonly="isSubmitted"
         :rules="[requiredRule]"
         item-title="value"
         item-value="key"
-        @update:model-value="selectedAnswer[premise.key] = $event"
+        variant="outlined"
+        @update:model-value="answer[premise.key] = $event"
       >
-        <template v-if="submitted" #append>
+        <template v-if="isSubmitted" #append>
           <VIcon v-bind="iconProps(premise.key)" />
         </template>
       </VSelect>
     </div>
-    <VAlert
-      v-if="submitted"
-      :text="userState?.isCorrect ? 'Correct' : 'Incorrect'"
-      :type="userState?.isCorrect ? 'success' : 'error'"
-      class="mb-3"
-      rounded="lg"
-      variant="tonal"
-    />
-    <div class="d-flex justify-end">
-      <VBtn v-if="!submitted" type="submit" variant="tonal">Submit</VBtn>
-      <VBtn v-else variant="tonal" @click="submitted = false">Try Again</VBtn>
-    </div>
-  </VForm>
+  </QuestionContainer>
 </template>
 
 <script setup lang="ts">
 import { ref, watch } from 'vue';
 import cloneDeep from 'lodash/cloneDeep';
 import { ElementData } from '@tailor-cms/ce-matching-question-manifest';
+import { QuestionContainer } from '@tailor-cms/lx-components';
 
 const initializeAnswer = () => cloneDeep(props.userState?.response) ?? {};
 
 const props = defineProps<{ id: number; data: ElementData; userState: any }>();
 const emit = defineEmits(['interaction']);
 
-const form = ref<HTMLFormElement>();
-const submitted = ref('isCorrect' in (props.userState ?? {}));
-const selectedAnswer = ref(initializeAnswer());
+const isSubmitted = ref(!!props.userState.isSubmitted);
+const answer = ref(initializeAnswer());
 
 const requiredRule = (val: string | boolean | number) => {
   return !!val || 'You have to select an answer.';
 };
 
-const submit = async () => {
-  const { valid } = await form.value?.validate();
-  if (valid) emit('interaction', { response: selectedAnswer.value });
-};
+const submit = () => emit('interaction', { response: answer.value });
 
 const iconProps = (uuid: string) => {
   const { response, correct } = props.userState;
@@ -68,8 +66,8 @@ const iconProps = (uuid: string) => {
 watch(
   () => props.userState,
   (state = {}) => {
-    selectedAnswer.value = initializeAnswer();
-    submitted.value = 'isCorrect' in state;
+    answer.value = initializeAnswer();
+    isSubmitted.value = !!state.isSubmitted;
   },
   { deep: true },
 );