+++ packages/server/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @tailor-cms/ce-matching-question-server
 
+## 0.0.6
+
+### Patch Changes
+
+- Bumped tce-display-runtime.
+
+## 0.0.5
+
+### Patch Changes
+
+- Refactored isGradable handling, by storing it directly in element data.
+
+## 0.0.4
+
+### Patch Changes
+
+- Refactored Edit and Display components to use QuestionContainer component from core-components/lx-components.
+
 ## 0.0.3
 
 ### Patch Changes